User Management Complete

+++ app/auth/__init__.py
@@ -12,8 +12,6 @@ auth = Blueprint('auth', __name__, template_folder='templates')
 def my_profile():
     user = User.query.get(current_user.get_id())
 
-
-
 @auth.route('/login', methods=['POST', 'GET'])
 def login():
     form = login_form()
@@ -74,30 +72,55 @@ def logout():
 @auth.route('/users/<int:user_id>')
 @login_required
 def retrieve_user(user_id):
-    pass
+    user = User.query.get(user_id)
+    return render_template('profile_view.html', user=user)
 
 
-@auth.route('/users/<int:user_id>/edit')
+@auth.route('/users/<int:user_id>/edit', methods=['POST', 'GET'])
 @login_required
 def edit_user(user_id):
-    pass
+    user = User.query.get(user_id)
+    form = register_form(obj=user)
+    if form.validate_on_submit():
+        user.email = form.email.data
+        user.password = form.password.data
+        db.session.add(user)
+        db.session.commit()
+        flash('User Edited Successfully')
+        return redirect(url_for('auth.browse_users'))
+    return render_template('profile_edit.html', form=form)
 
 
 @auth.route('/users/<int:user_id>/delete', methods=['POST'])
 @login_required
 def delete_user(user_id):
     user = User.query.get(user_id)
+    logged_in_user = current_user
+    if user.id == logged_in_user.id:
+        flash("You can't delete yourself!")
+        return redirect(url_for('auth.browse_users'), 302)
     db.session.delete(user)
     db.session.commit()
     flash('User Deleted')
-    return redirect('auth.browse_users', 302)
+    return redirect(url_for('auth.browse_users'), 302)
 
 
-@auth.route('/users/new')
+@auth.route('/users/new', methods=['POST', 'GET'])
 @login_required
 def add_user():
-    pass
-
+    form = register_form()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first()
+        if user is None:
+            user = User(email=form.email.data, password=generate_password_hash(form.password.data))
+            db.session.add(user)
+            db.session.commit()
+            flash('Congratulations, you just created a user')
+            return redirect(url_for('auth.browse_users'))
+        else:
+            flash('Already Registered')
+            return redirect(url_for('auth.browse_users'))
+    return render_template('profile_new.html', form=form)
 
 @auth.route('/users')
 @login_required
